change card columns to masonry, format buttons

+++ app/Controllers/ListController.js
@@ -47,9 +47,9 @@ export default class ListController {
   }
 
   deleteList(listId) {
-    // if (!window.confirm("Delete list?")) {
-    //   return;
-    // }
+    if (!window.confirm("Delete list?")) {
+      return;
+    }
     ListService.deleteList(listId);
     _drawLists();
   }
